init litellm skills handler
