+++ litellm/llms/litellm_proxy/skills/__init__.py
@@ -5,11 +5,13 @@ This module provides database-backed skills storage as an alternative to
 Anthropic's cloud-based skills API.
 
 Main components:
-- handler.py: LiteLLMSkillsTransformationHandler - sync/async methods for CRUD
-- transformation.py: LiteLLMSkillsConfig - BaseSkillsAPIConfig implementation (optional)
+- handler.py: LiteLLMSkillsHandler - database CRUD operations
+- transformation.py: LiteLLMSkillsTransformationHandler - SDK transformation layer
 """
 
-from litellm.llms.litellm_proxy.skills.handler import (
+from litellm.llms.litellm_proxy.skills.handler import LiteLLMSkillsHandler
+from litellm.llms.litellm_proxy.skills.transformation import (
     LiteLLMSkillsTransformationHandler,
 )
 
+__all__ = ["LiteLLMSkillsHandler", "LiteLLMSkillsTransformationHandler"]
